refactor(clippings): merge recipes workspace into clippings
Recipes are conceptually a type of "clipping" (saved content), so they belong
in the clippings workspace alongside articles, books, and github repos.

This consolidates the recipes table, custom markdown serialization (with
## Ingredients and ## Instructions sections), and mutations (addRecipe,
updateRecipeRating) into clippings.workspace.ts. The example script has been
moved to clippings/scripts/ and now uses client.clippings.addRecipe().
